+++ kartothek/io_components/metapartition.py
@@ -620,8 +620,7 @@ class MetaPartition(Iterable):
         dates_as_object: bool
             Load pyarrow.data{32,64} columns as ``object`` columns in Pandas
             instead of using ``np.datetime64`` to preserve their type. While
-            this improves type-safety, this comes at a performance cost. Only
-            works for metadata version >= 4.
+            this improves type-safety, this comes at a performance cost.
 
         Returns
         -------